Bug Fixes/More
When the sidebar is removed from the grid on destroy, it sends a combo box selection changed event but the item is null by the time it processes it.

Worked on the functionality that allows users to select appointments, modify them and return home when cancelling or saving changes.

Added fields and properties to the Appointment class to allow for pre selecting the active appt's values when the modify appointment sidebar is first opened.